增加了新的數據對象設計，增加了文章和圖片保存的service 和controller

+++ src/main/java/com/miwo/vo/InheritVo.java
@@ -10,7 +10,7 @@ public class InheritVo {
 	 *
 	 * @mbg.generated Wed Apr 25 10:19:05 CST 2018
 	 */
-	private Integer id;
+	private Long id;
 
 	/**
 	 *
@@ -39,11 +39,11 @@ public class InheritVo {
 	 */
 	private List<String> children;
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 
@@ -71,11 +71,11 @@ public class InheritVo {
 		this.children = children;
 	}
 
-	public Integer getUserId() {
+	public Long getUserId() {
 		return userId;
 	}
 
-	public void setUserId(Integer userId) {
+	public void setUserId(Long userId) {
 		this.userId = userId;
 	}
 
@@ -86,5 +86,5 @@ public class InheritVo {
 	 *
 	 * @mbg.generated Wed Apr 25 10:19:05 CST 2018
 	 */
-	private Integer userId;
+	private Long userId;
 }
